Reorganise + cleanup some code
