+++ packages/apps/tools/.eslintrc.js
@@ -12,6 +12,7 @@ module.exports = {
     '@kadena-dev/typedef-var': 'off',
     '@typescript-eslint/explicit-function-return-type': 'off',
     '@rushstack/no-new-null': 'off',
+    '@typescript-eslint/strict-boolean-expressions': 'off',
     'no-void': 'off',
     'react-hooks/rules-of-hooks': 'error',
     'react-hooks/exhaustive-deps': [
