+++ packages/apps/tools/src/components/Global/ModuleExplorer/index.tsx
@@ -1,7 +1,5 @@
 import type { IncompleteModuleModel } from '@/pages/modules/explorer/utils';
 import { isModule } from '@/pages/modules/explorer/utils';
-import type { ChainwebChainId } from '@kadena/chainweb-node-client';
-import { contractParser } from '@kadena/pactjs-generator';
 import React, { useState } from 'react';
 import type { TreeItem } from '../CustomTree/CustomTree';
 import type { ISidePanelProps } from './SidePanel';
@@ -9,7 +7,8 @@ import SidePanel from './SidePanel';
 import type { IEditorProps } from './editor';
 import Editor from './editor';
 import { containerStyle } from './styles.css';
-import type { IChainModule } from './types';
+import type { IChainModule, Outline } from './types';
+import { chainModuleToOutlineTreeItems } from './utils';
 
 export interface IModuleExplorerProps {
   openedModules: IEditorProps['openedModules'];
@@ -22,100 +21,6 @@ export interface IModuleExplorerProps {
     IncompleteModuleModel | Outline
   >['onExpandCollapse'];
 }
-
-type ElementType<T> = T extends (infer U)[] ? U : never;
-
-export type Contract = ReturnType<typeof contractParser>[0][0]; // TODO: Should we improve this because it's a bit hacky?
-export type ContractInterface = ElementType<Contract['usedInterface']> & {
-  code?: string;
-  chainId: ChainwebChainId;
-  networkId: string;
-};
-export type ContractCapability = ElementType<Contract['capabilities']>;
-export type ContractFunction = ElementType<Contract['functions']>;
-
-export type Outline =
-  | string
-  | ContractInterface
-  | ContractCapability
-  | ContractFunction;
-
-const chainModuleToOutlineTreeItems = (
-  chainModule: IChainModule,
-  items: TreeItem<IncompleteModuleModel>[],
-): TreeItem<Outline>[] => {
-  const treeItems: TreeItem<Outline>[] = [];
-
-  if (!chainModule.code) {
-    return treeItems;
-  }
-
-  const [, namespace] = chainModule.moduleName.split('.');
-  const [[parsedContract]] = contractParser(chainModule.code, namespace);
-
-  const { usedInterface: interfaces, capabilities, functions } = parsedContract;
-
-  if (interfaces?.length) {
-    treeItems.push({
-      title: 'Interfaces',
-      key: 'interfaces',
-      data: 'interfaces',
-      children: interfaces.map((i) => {
-        const firstFind = items.find((item) => {
-          return item.data.name === chainModule.network;
-        });
-        const secondFind = firstFind?.children.find((child) => {
-          return child.data.name === i.name;
-        });
-        const thirdFind = secondFind?.children.find((child) => {
-          return child.data.chainId === chainModule.chainId;
-        });
-        return {
-          title: i.name,
-          key: `${chainModule.network}.${i.name}`,
-          label: thirdFind?.data.hash,
-          data: {
-            ...i,
-            chainId: chainModule.chainId,
-            networkId: chainModule.network,
-            code: thirdFind?.data.code,
-          },
-          children: [],
-        };
-      }),
-    });
-  }
-
-  if (capabilities?.length) {
-    treeItems.push({
-      title: 'Capabilities',
-      key: 'capabilities',
-      data: 'capabilities',
-      children: capabilities.map((c) => ({
-        title: c.name,
-        key: c.name,
-        data: c,
-        children: [],
-      })),
-    });
-  }
-
-  if (functions?.length) {
-    treeItems.push({
-      title: 'Functions',
-      key: 'functions',
-      data: 'functions',
-      children: functions.map((f) => ({
-        title: f.name,
-        key: f.name,
-        data: f,
-        children: [],
-      })),
-    });
-  }
-
-  return treeItems;
-};
 
 const ModuleExplorer = ({
   items,
