+++ packages/apps/tools/src/components/Global/ModuleExplorer/types.ts
@@ -1,5 +1,10 @@
 import type { Network } from '@/constants/kadena';
 import type { ChainwebChainId } from '@kadena/chainweb-node-client';
+import type {
+  ContractCapability,
+  ContractFunction,
+  ContractInterface,
+} from './utils';
 
 export interface IChainModule {
   code?: string;
@@ -13,3 +18,11 @@ export interface IModule {
   moduleName: string;
   chains: ChainwebChainId[];
 }
+
+export type ElementType<T> = T extends (infer U)[] ? U : never;
+
+export type Outline =
+  | string
+  | ContractInterface
+  | ContractCapability
+  | ContractFunction;
