+++ packages/apps/tools/src/components/Global/ModuleExplorer/utils.ts
@@ -1,50 +1,91 @@
-import type { Network } from '@/constants/kadena';
+import type { IncompleteModuleModel } from '@/pages/modules/explorer/utils';
 import type { ChainwebChainId } from '@kadena/chainweb-node-client';
-import type { IChainModule } from './types';
+import { contractParser } from '@kadena/pactjs-generator';
+import type { TreeItem } from '../CustomTree/CustomTree';
+import type { ElementType, IChainModule, Outline } from './types';
 
-export const getModulesMap = (
-  modules: IChainModule[],
-): Map<
-  string,
-  Array<{
-    chainId: ChainwebChainId;
-    hash?: string;
-    code?: string;
-    network: Network;
-  }>
-> => {
-  const modulesMap = new Map<
-    string,
-    Array<{
-      chainId: ChainwebChainId;
-      hash?: string;
-      code?: string;
-      network: Network;
-    }>
-  >();
+export type Contract = ReturnType<typeof contractParser>[0][0]; // TODO: Should we improve this because it's a bit hacky?
+export type ContractInterface = ElementType<Contract['usedInterface']> & {
+  code?: string;
+  chainId: ChainwebChainId;
+  networkId: string;
+};
+export type ContractCapability = ElementType<Contract['capabilities']>;
+export type ContractFunction = ElementType<Contract['functions']>;
 
-  modules.forEach((module) => {
-    if (modulesMap.has(module.moduleName)) {
-      modulesMap.set(module.moduleName, [
-        ...modulesMap.get(module.moduleName)!,
-        {
-          chainId: module.chainId,
-          hash: module.hash,
-          code: module.code,
-          network: module.network,
-        },
-      ]);
-    } else {
-      modulesMap.set(module.moduleName, [
-        {
-          chainId: module.chainId,
-          hash: module.hash,
-          code: module.code,
-          network: module.network,
-        },
-      ]);
-    }
-  });
+export const chainModuleToOutlineTreeItems = (
+  chainModule: IChainModule,
+  items: TreeItem<IncompleteModuleModel>[],
+): TreeItem<Outline>[] => {
+  const treeItems: TreeItem<Outline>[] = [];
 
-  return modulesMap;
+  if (!chainModule.code) {
+    return treeItems;
+  }
+
+  const [, namespace] = chainModule.moduleName.split('.');
+  const [[parsedContract]] = contractParser(chainModule.code, namespace);
+
+  const { usedInterface: interfaces, capabilities, functions } = parsedContract;
+
+  if (interfaces?.length) {
+    treeItems.push({
+      title: 'Interfaces',
+      key: 'interfaces',
+      data: 'interfaces',
+      children: interfaces.map((i) => {
+        const firstFind = items.find((item) => {
+          return item.data.name === chainModule.network;
+        });
+        const secondFind = firstFind?.children.find((child) => {
+          return child.data.name === i.name;
+        });
+        const thirdFind = secondFind?.children.find((child) => {
+          return child.data.chainId === chainModule.chainId;
+        });
+        return {
+          title: i.name,
+          key: `${chainModule.network}.${i.name}`,
+          label: thirdFind?.data.hash,
+          data: {
+            ...i,
+            chainId: chainModule.chainId,
+            networkId: chainModule.network,
+            code: thirdFind?.data.code,
+          },
+          children: [],
+        };
+      }),
+    });
+  }
+
+  if (capabilities?.length) {
+    treeItems.push({
+      title: 'Capabilities',
+      key: 'capabilities',
+      data: 'capabilities',
+      children: capabilities.map((c) => ({
+        title: c.name,
+        key: c.name,
+        data: c,
+        children: [],
+      })),
+    });
+  }
+
+  if (functions?.length) {
+    treeItems.push({
+      title: 'Functions',
+      key: 'functions',
+      data: 'functions',
+      children: functions.map((f) => ({
+        title: f.name,
+        key: f.name,
+        data: f,
+        children: [],
+      })),
+    });
+  }
+
+  return treeItems;
 };
